Features: dependencies optimisation & adds bundle-analyzer
This commit adds the bundle-analyzer plugin that will run whenever
the build command is run. This makes optimising the app's size easier
as that it will be possible to see what is taking the majority of the
size.

This commit also removes the dependency on polished which is highly
opinionated and not needed for a lot of use cases, an there is
@theme-ui/color which does the same thing and works better with
theme-ui.